work: consolidate tests into lib/work/
- Move test files from lib/test/work/ to lib/work/
- Rename with test_ prefix for clarity
- Consolidate runner into lib/work/test.lua
- Update requires to use work.test_lib
