+++ lib/work/test_file_locking.lua
@@ -3,7 +3,7 @@ local lu = require("luaunit")
 local data = require("work.data")
 local store = require("work.store")
 local fcntl = require("posix.fcntl")
-local Work = require("lib")
+local Work = require("work.test_lib")
 local test_store = Work.store
 
 TestFileLocking = {}
