+++ lib/work/test_string_sanitization.lua
@@ -2,7 +2,7 @@ local lu = require("luaunit")
 
 local data = require("work.data")
 local store = require("work.store")
-local Work = require("lib")
+local Work = require("work.test_lib")
 local test_store = Work.store
 
 TestStringSanitization = {}
